hexagon: move ops context into main context
Just a cleanup. We don't need separate contexts at this point.
